backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pydantic
pydantic-settings
pandas
gspread
google-auth
google-api-python-client
Pillow
python-multipart
google-auth-oauthlib
google-auth-httplib2
httpx
sqlalchemy
psycopg2-binary
pytest
slowapi
pytest-asyncio
